switch internal field to int64

+++ pkg/apis/extensions/types.go
@@ -333,7 +333,7 @@ type RollingUpdateDeployment struct {
 
 type DeploymentStatus struct {
 	// The generation observed by the deployment controller.
-	ObservedGeneration int `json:"observedGeneration,omitempty"`
+	ObservedGeneration int64 `json:"observedGeneration,omitempty"`
 
 	// Total number of non-terminated pods targeted by this deployment (their labels match the selector).
 	Replicas int `json:"replicas,omitempty"`
